+++ packages/next/src/server/lib/router-utils/setup-dev-bundler.ts
@@ -1,36 +1,11 @@
 import type { NextConfigComplete } from '../../config-shared'
-import type {
-  Endpoint,
-  Route,
-  TurbopackResult,
-  WrittenEndpoint,
-  Issue,
-  StyledString,
-} from '../../../build/swc'
-import type { Socket } from 'net'
 import type { FilesystemDynamicRoute } from './filesystem'
 import type { UnwrapPromise } from '../../../lib/coalesced-function'
 import type { MiddlewareMatcher } from '../../../build/analysis/get-page-static-info'
-import type { OutputState } from '../../../build/output/store'
 import type { MiddlewareRouteMatch } from '../../../shared/lib/router/utils/middleware-route-matcher'
-import type { BuildManifest } from '../../get-page-files'
-import type { PagesManifest } from '../../../build/webpack/plugins/pages-manifest-plugin'
-import type { AppBuildManifest } from '../../../build/webpack/plugins/app-build-manifest-plugin'
 import type { PropagateToWorkersField } from './types'
-import type {
-  EdgeFunctionDefinition,
-  MiddlewareManifest,
-} from '../../../build/webpack/plugins/middleware-plugin'
-import type {
-  CompilationError,
-  HMR_ACTION_TYPES,
-  NextJsHotReloaderInterface,
-  ReloadPageAction,
-  SyncAction,
-  TurbopackConnectedAction,
-} from '../../dev/hot-reloader-types'
+import type { NextJsHotReloaderInterface } from '../../dev/hot-reloader-types'
 
-import ws from 'next/dist/compiled/ws'
 import { createDefineEnv } from '../../../build/swc'
 import fs from 'fs'
 import url from 'url'
@@ -42,10 +17,7 @@ import isError from '../../../lib/is-error'
 import findUp from 'next/dist/compiled/find-up'
 import { buildCustomRoute } from './filesystem'
 import * as Log from '../../../build/output/log'
-import HotReloaderWebpack, {
-  getVersionInfo,
-  matchNextPageBundleRequest,
-} from '../../dev/hot-reloader-webpack'
+import HotReloaderWebpack from '../../dev/hot-reloader-webpack'
 import { setGlobal } from '../../../trace/shared'
 import type { Telemetry } from '../../../telemetry/storage'
 import type { IncomingMessage, ServerResponse } from 'http'
@@ -68,29 +40,14 @@ import { getRouteMatcher } from '../../../shared/lib/router/utils/route-matcher'
 import { normalizePathSep } from '../../../shared/lib/page-path/normalize-path-sep'
 import { createClientRouterFilter } from '../../../lib/create-client-router-filter'
 import { absolutePathToPage } from '../../../shared/lib/page-path/absolute-path-to-page'
-import {
-  generateInterceptionRoutesRewrites,
-  isInterceptionRouteRewrite,
-} from '../../../lib/generate-interception-routes-rewrites'
-import { store as consoleStore } from '../../../build/output/store'
+import { generateInterceptionRoutesRewrites } from '../../../lib/generate-interception-routes-rewrites'
 
 import {
-  APP_BUILD_MANIFEST,
-  APP_PATHS_MANIFEST,
-  BUILD_MANIFEST,
   CLIENT_STATIC_FILES_PATH,
   COMPILER_NAMES,
   DEV_CLIENT_PAGES_MANIFEST,
   DEV_MIDDLEWARE_MANIFEST,
-  MIDDLEWARE_MANIFEST,
-  NEXT_FONT_MANIFEST,
-  PAGES_MANIFEST,
   PHASE_DEVELOPMENT_SERVER,
-  SERVER_REFERENCE_MANIFEST,
-  REACT_LOADABLE_MANIFEST,
-  MIDDLEWARE_REACT_LOADABLE_MANIFEST,
-  MIDDLEWARE_BUILD_MANIFEST,
-  INTERCEPTION_ROUTE_REWRITE_MANIFEST,
 } from '../../../shared/lib/constants'
 
 import { getMiddlewareRouteMatcher } from '../../../shared/lib/router/utils/middleware-route-matcher'
@@ -108,63 +65,14 @@ import {
   getSourceById,
   parseStack,
 } from 'next/dist/compiled/@next/react-dev-overlay/dist/middleware'
-import {
-  getOverlayMiddleware,
-  createOriginalStackFrame as createOriginalTurboStackFrame,
-} from 'next/dist/compiled/@next/react-dev-overlay/dist/middleware-turbopack'
-import { mkdir, readFile, writeFile } from 'fs/promises'
-import { PageNotFoundError } from '../../../shared/lib/utils'
-import {
-  type ClientBuildManifest,
-  normalizeRewritesForBuildManifest,
-  srcEmptySsgManifest,
-} from '../../../build/webpack/plugins/build-manifest-plugin'
+import { createOriginalStackFrame as createOriginalTurboStackFrame } from 'next/dist/compiled/@next/react-dev-overlay/dist/middleware-turbopack'
 import { devPageFiles } from '../../../build/webpack/plugins/next-types-plugin/shared'
 import type { LazyRenderServerInstance } from '../router-server'
-import { pathToRegexp } from 'next/dist/compiled/path-to-regexp'
 import { HMR_ACTIONS_SENT_TO_BROWSER } from '../../dev/hot-reloader-types'
-import type { Update as TurbopackUpdate } from '../../../build/swc'
-import { debounce } from '../../utils'
-import {
-  deleteAppClientCache,
-  deleteCache,
-} from '../../../build/webpack/plugins/nextjs-require-cache-hot-reloader'
-import { normalizeMetadataRoute } from '../../../lib/metadata/get-metadata-route'
-import { clearModuleContext, clearAllModuleContexts } from '../render-server'
-import type { ActionManifest } from '../../../build/webpack/plugins/flight-client-entry-plugin'
-import { denormalizePagePath } from '../../../shared/lib/page-path/denormalize-page-path'
-import type { LoadableManifest } from '../../load-components'
-import { generateRandomActionKeyRaw } from '../../app-render/action-encryption-utils'
-import { bold, green, magenta, red } from '../../../lib/picocolors'
-import { writeFileAtomic } from '../../../lib/fs/write-atomic'
 import { PAGE_TYPES } from '../../../lib/page-types'
-import { trace } from '../../../trace'
-import type { VersionInfo } from '../../dev/parse-version-info'
-import type { NextFontManifest } from '../../../build/webpack/plugins/next-font-manifest-plugin'
-import {
-  MAGIC_IDENTIFIER_REGEX,
-  decodeMagicIdentifier,
-} from '../../../shared/lib/magic-identifier'
-import {
-  mergeActionManifests,
-  mergeAppBuildManifests,
-  mergeBuildManifests,
-  mergeFontManifests,
-  mergeLoadableManifests,
-  mergeMiddlewareManifests,
-  mergePagesManifests,
-  type TurbopackMiddlewareManifest,
-} from './turbopack-utils'
+import { createHotReloaderTurbopack } from '../../dev/hot-reloader-turbopack'
 
-const MILLISECONDS_IN_NANOSECOND = 1_000_000
-const wsServer = new ws.Server({ noServer: true })
-const isTestMode = !!(
-  process.env.NEXT_TEST_MODE ||
-  process.env.__NEXT_TEST_MODE ||
-  process.env.DEBUG
-)
-
-type SetupOpts = {
+export type SetupOpts = {
   renderServer: LazyRenderServerInstance
   dir: string
   turbo?: boolean
@@ -177,6 +85,23 @@ type SetupOpts = {
   >
   nextConfig: NextConfigComplete
   port: number
+}
+
+export type ServerFields = {
+  actualMiddlewareFile?: string | undefined
+  actualInstrumentationHookFile?: string | undefined
+  appPathRoutes?: Record<string, string | string[]>
+  middleware?:
+    | {
+        page: string
+        match: MiddlewareRouteMatch
+        matchers?: MiddlewareMatcher[]
+      }
+    | undefined
+  hasAppNotFound?: boolean
+  interceptionRoutes?: ReturnType<
+    typeof import('./filesystem').buildCustomRoute
+  >[]
 }
 
 async function verifyTypeScript(opts: SetupOpts) {
@@ -200,6 +125,14 @@ async function verifyTypeScript(opts: SetupOpts) {
 
 class ModuleBuildError extends Error {}
 
+export async function propagateServerField(
+  opts: SetupOpts,
+  field: PropagateToWorkersField,
+  args: any
+) {
+  await opts.renderServer?.instance?.propagateServerField(opts.dir, field, args)
+}
+
 async function startWatcher(opts: SetupOpts) {
   const { nextConfig, appDir, pagesDir, dir } = opts
   const { useFileSystemPublicRoutes } = nextConfig
@@ -215,1475 +148,10 @@ async function startWatcher(opts: SetupOpts) {
     appDir
   )
 
-  async function propagateServerField(
-    field: PropagateToWorkersField,
-    args: any
-  ) {
-    await opts.renderServer?.instance?.propagateServerField(
-      opts.dir,
-      field,
-      args
-    )
-  }
-
-  const serverFields: {
-    actualMiddlewareFile?: string | undefined
-    actualInstrumentationHookFile?: string | undefined
-    appPathRoutes?: Record<string, string | string[]>
-    middleware?:
-      | {
-          page: string
-          match: MiddlewareRouteMatch
-          matchers?: MiddlewareMatcher[]
-        }
-      | undefined
-    hasAppNotFound?: boolean
-    interceptionRoutes?: ReturnType<
-      typeof import('./filesystem').buildCustomRoute
-    >[]
-  } = {}
-
-  async function createHotReloaderTurbopack(): Promise<NextJsHotReloaderInterface> {
-    const { loadBindings } =
-      require('../../../build/swc') as typeof import('../../../build/swc')
-
-    let bindings = await loadBindings()
-
-    const { jsConfig } = await loadJsConfig(dir, opts.nextConfig)
-
-    // For the debugging purpose, check if createNext or equivalent next instance setup in test cases
-    // works correctly. Normally `run-test` hides output so only will be visible when `--debug` flag is used.
-    if (process.env.TURBOPACK && isTestMode) {
-      require('console').log('Creating turbopack project', {
-        dir,
-        testMode: isTestMode,
-      })
-    }
-
-    const hasRewrites =
-      opts.fsChecker.rewrites.afterFiles.length > 0 ||
-      opts.fsChecker.rewrites.beforeFiles.length > 0 ||
-      opts.fsChecker.rewrites.fallback.length > 0
-
-    const hotReloaderSpan = trace('hot-reloader', undefined, {
-      version: process.env.__NEXT_VERSION as string,
-    })
-    // Ensure the hotReloaderSpan is flushed immediately as it's the parentSpan for all processing
-    // of the current `next dev` invocation.
-    hotReloaderSpan.stop()
-
-    const project = await bindings.turbo.createProject({
-      projectPath: dir,
-      rootPath: opts.nextConfig.experimental.outputFileTracingRoot || dir,
-      nextConfig: opts.nextConfig,
-      jsConfig: jsConfig ?? { compilerOptions: {} },
-      watch: true,
-      env: process.env as Record<string, string>,
-      defineEnv: createDefineEnv({
-        isTurbopack: true,
-        allowedRevalidateHeaderKeys: undefined,
-        clientRouterFilters: undefined,
-        config: nextConfig,
-        dev: true,
-        distDir,
-        fetchCacheKeyPrefix: undefined,
-        hasRewrites,
-        middlewareMatchers: undefined,
-        previewModeId: undefined,
-      }),
-      serverAddr: `127.0.0.1:${opts.port}`,
-    })
-    const iter = project.entrypointsSubscribe()
-    const curEntries: Map<string, Route> = new Map()
-    const changeSubscriptions: Map<
-      string,
-      Promise<AsyncIterator<any>>
-    > = new Map()
-    let prevMiddleware: boolean | undefined = undefined
-    const globalEntries: {
-      app: Endpoint | undefined
-      document: Endpoint | undefined
-      error: Endpoint | undefined
-    } = {
-      app: undefined,
-      document: undefined,
-      error: undefined,
-    }
-    let currentEntriesHandlingResolve: ((value?: unknown) => void) | undefined
-    let currentEntriesHandling = new Promise(
-      (resolve) => (currentEntriesHandlingResolve = resolve)
-    )
-    const hmrPayloads = new Map<string, HMR_ACTION_TYPES>()
-    const turbopackUpdates: TurbopackUpdate[] = []
-
-    const issues = new Map<string, Map<string, Issue>>()
-
-    function issueKey(issue: Issue): string {
-      return [
-        issue.severity,
-        issue.filePath,
-        JSON.stringify(issue.title),
-        JSON.stringify(issue.description),
-      ].join('-')
-    }
-
-    function formatIssue(issue: Issue) {
-      const { filePath, title, description, source } = issue
-      let { documentationLink } = issue
-      let formattedTitle = renderStyledStringToErrorAnsi(title).replace(
-        /\n/g,
-        '\n    '
-      )
-
-      // TODO: Use error codes to identify these
-      // TODO: Generalize adapting Turbopack errors to Next.js errors
-      if (formattedTitle.includes('Module not found')) {
-        // For compatiblity with webpack
-        // TODO: include columns in webpack errors.
-        documentationLink = 'https://nextjs.org/docs/messages/module-not-found'
-      }
-
-      let formattedFilePath = filePath
-        .replace('[project]/', './')
-        .replaceAll('/./', '/')
-        .replace('\\\\?\\', '')
-
-      let message
-
-      if (source && source.range) {
-        const { start } = source.range
-        message = `${formattedFilePath}:${start.line + 1}:${
-          start.column + 1
-        }\n${formattedTitle}`
-      } else if (formattedFilePath) {
-        message = `${formattedFilePath}\n${formattedTitle}`
-      } else {
-        message = formattedTitle
-      }
-      message += '\n'
-
-      if (source?.range && source.source.content) {
-        const { start, end } = source.range
-        const {
-          codeFrameColumns,
-        } = require('next/dist/compiled/babel/code-frame')
-
-        message +=
-          codeFrameColumns(
-            source.source.content,
-            {
-              start: {
-                line: start.line + 1,
-                column: start.column + 1,
-              },
-              end: {
-                line: end.line + 1,
-                column: end.column + 1,
-              },
-            },
-            { forceColor: true }
-          ).trim() + '\n\n'
-      }
-
-      if (description) {
-        message += renderStyledStringToErrorAnsi(description) + '\n\n'
-      }
-
-      // TODO: make it possible to enable this for debugging, but not in tests.
-      // if (detail) {
-      //   message += renderStyledStringToErrorAnsi(detail) + '\n\n'
-      // }
-
-      // TODO: Include a trace from the issue.
-
-      if (documentationLink) {
-        message += documentationLink + '\n\n'
-      }
-
-      return message
-    }
-
-    function processIssues(
-      name: string,
-      result: TurbopackResult,
-      throwIssue = false
-    ) {
-      const newIssues = new Map<string, Issue>()
-      issues.set(name, newIssues)
-
-      const relevantIssues = new Set()
-
-      for (const issue of result.issues) {
-        if (issue.severity !== 'error' && issue.severity !== 'fatal') continue
-        const key = issueKey(issue)
-        const formatted = formatIssue(issue)
-        newIssues.set(key, issue)
-
-        // We show errors in node_modules to the console, but don't throw for them
-        if (/(^|\/)node_modules(\/|$)/.test(issue.filePath)) continue
-        relevantIssues.add(formatted)
-      }
-
-      if (relevantIssues.size && throwIssue) {
-        throw new ModuleBuildError([...relevantIssues].join('\n\n'))
-      }
-    }
-
-    const serverPathState = new Map<string, string>()
-
-    async function processResult(
-      id: string,
-      result: TurbopackResult<WrittenEndpoint>
-    ): Promise<TurbopackResult<WrittenEndpoint>> {
-      // Figure out if the server files have changed
-      let hasChange = false
-      for (const { path: p, contentHash } of result.serverPaths) {
-        // We ignore source maps
-        if (p.endsWith('.map')) continue
-        let key = `${id}:${p}`
-        const localHash = serverPathState.get(key)
-        const globalHash = serverPathState.get(p)
-        if (
-          (localHash && localHash !== contentHash) ||
-          (globalHash && globalHash !== contentHash)
-        ) {
-          hasChange = true
-          serverPathState.set(key, contentHash)
-          serverPathState.set(p, contentHash)
-        } else {
-          if (!localHash) {
-            serverPathState.set(key, contentHash)
-          }
-          if (!globalHash) {
-            serverPathState.set(p, contentHash)
-          }
-        }
-      }
-
-      if (!hasChange) {
-        return result
-      }
-
-      const hasAppPaths = result.serverPaths.some(({ path: p }) =>
-        p.startsWith('server/app')
-      )
-
-      if (hasAppPaths) {
-        deleteAppClientCache()
-      }
-
-      const serverPaths = result.serverPaths.map(({ path: p }) =>
-        path.join(distDir, p)
-      )
-
-      for (const file of serverPaths) {
-        clearModuleContext(file)
-        deleteCache(file)
-      }
-
-      return result
-    }
-
-    const buildingIds = new Set()
-    const readyIds = new Set()
-
-    function startBuilding(
-      id: string,
-      requestUrl: string | undefined,
-      forceRebuild: boolean = false
-    ) {
-      if (!forceRebuild && readyIds.has(id)) {
-        return () => {}
-      }
-      if (buildingIds.size === 0) {
-        consoleStore.setState(
-          {
-            loading: true,
-            trigger: id,
-            url: requestUrl,
-          } as OutputState,
-          true
-        )
-      }
-      buildingIds.add(id)
-      return function finishBuilding() {
-        if (buildingIds.size === 0) {
-          return
-        }
-        readyIds.add(id)
-        buildingIds.delete(id)
-        if (buildingIds.size === 0) {
-          consoleStore.setState(
-            {
-              loading: false,
-            } as OutputState,
-            true
-          )
-        }
-      }
-    }
-
-    let hmrEventHappened = false
-    let hmrHash = 0
-    const sendEnqueuedMessages = () => {
-      for (const [, issueMap] of issues) {
-        if (issueMap.size > 0) {
-          // During compilation errors we want to delay the HMR events until errors are fixed
-          return
-        }
-      }
-      for (const payload of hmrPayloads.values()) {
-        hotReloader.send(payload)
-      }
-      hmrPayloads.clear()
-      if (turbopackUpdates.length > 0) {
-        hotReloader.send({
-          action: HMR_ACTIONS_SENT_TO_BROWSER.TURBOPACK_MESSAGE,
-          data: turbopackUpdates,
-        })
-        turbopackUpdates.length = 0
-      }
-    }
-    const sendEnqueuedMessagesDebounce = debounce(sendEnqueuedMessages, 2)
-
-    function sendHmr(key: string, id: string, payload: HMR_ACTION_TYPES) {
-      hmrPayloads.set(`${key}:${id}`, payload)
-      hmrEventHappened = true
-      sendEnqueuedMessagesDebounce()
-    }
-
-    function sendTurbopackMessage(payload: TurbopackUpdate) {
-      turbopackUpdates.push(payload)
-      hmrEventHappened = true
-      sendEnqueuedMessagesDebounce()
-    }
-
-    async function loadPartialManifest<T>(
-      name: string,
-      pageName: string,
-      type:
-        | 'pages'
-        | 'app'
-        | 'app-route'
-        | 'middleware'
-        | 'instrumentation' = 'pages'
-    ): Promise<T> {
-      const manifestPath = path.posix.join(
-        distDir,
-        `server`,
-        type === 'app-route' ? 'app' : type,
-        type === 'middleware' || type === 'instrumentation'
-          ? ''
-          : pageName === '/'
-          ? 'index'
-          : pageName === '/index' || pageName.startsWith('/index/')
-          ? `/index${pageName}`
-          : pageName,
-        type === 'app' ? 'page' : type === 'app-route' ? 'route' : '',
-        name
-      )
-      return JSON.parse(
-        await readFile(path.posix.join(manifestPath), 'utf-8')
-      ) as T
-    }
-
-    const buildManifests = new Map<string, BuildManifest>()
-    const appBuildManifests = new Map<string, AppBuildManifest>()
-    const pagesManifests = new Map<string, PagesManifest>()
-    const appPathsManifests = new Map<string, PagesManifest>()
-    const middlewareManifests = new Map<string, TurbopackMiddlewareManifest>()
-    const actionManifests = new Map<string, ActionManifest>()
-    const fontManifests = new Map<string, NextFontManifest>()
-    const loadableManifests = new Map<string, LoadableManifest>()
-    const clientToHmrSubscription = new Map<
-      ws,
-      Map<string, AsyncIterator<any>>
-    >()
-    const clients = new Set<ws>()
-
-    async function loadMiddlewareManifest(
-      pageName: string,
-      type: 'pages' | 'app' | 'app-route' | 'middleware' | 'instrumentation'
-    ): Promise<void> {
-      middlewareManifests.set(
-        pageName,
-        await loadPartialManifest(MIDDLEWARE_MANIFEST, pageName, type)
-      )
-    }
-
-    async function loadBuildManifest(
-      pageName: string,
-      type: 'app' | 'pages' = 'pages'
-    ): Promise<void> {
-      buildManifests.set(
-        pageName,
-        await loadPartialManifest(BUILD_MANIFEST, pageName, type)
-      )
-    }
-
-    async function loadAppBuildManifest(pageName: string): Promise<void> {
-      appBuildManifests.set(
-        pageName,
-        await loadPartialManifest(APP_BUILD_MANIFEST, pageName, 'app')
-      )
-    }
-
-    async function loadPagesManifest(pageName: string): Promise<void> {
-      pagesManifests.set(
-        pageName,
-        await loadPartialManifest(PAGES_MANIFEST, pageName)
-      )
-    }
-
-    async function loadAppPathManifest(
-      pageName: string,
-      type: 'app' | 'app-route' = 'app'
-    ): Promise<void> {
-      appPathsManifests.set(
-        pageName,
-        await loadPartialManifest(APP_PATHS_MANIFEST, pageName, type)
-      )
-    }
-
-    async function loadActionManifest(pageName: string): Promise<void> {
-      actionManifests.set(
-        pageName,
-        await loadPartialManifest(
-          `${SERVER_REFERENCE_MANIFEST}.json`,
-          pageName,
-          'app'
-        )
-      )
-    }
-
-    async function loadFontManifest(
-      pageName: string,
-      type: 'app' | 'pages' = 'pages'
-    ): Promise<void> {
-      fontManifests.set(
-        pageName,
-        await loadPartialManifest(`${NEXT_FONT_MANIFEST}.json`, pageName, type)
-      )
-    }
-
-    async function loadLoadableManifest(
-      pageName: string,
-      type: 'app' | 'pages' = 'pages'
-    ): Promise<void> {
-      loadableManifests.set(
-        pageName,
-        await loadPartialManifest(REACT_LOADABLE_MANIFEST, pageName, type)
-      )
-    }
-
-    async function changeSubscription(
-      page: string,
-      type: 'client' | 'server',
-      includeIssues: boolean,
-      endpoint: Endpoint | undefined,
-      makePayload: (
-        page: string,
-        change: TurbopackResult
-      ) => Promise<HMR_ACTION_TYPES> | HMR_ACTION_TYPES | void
-    ) {
-      const key = `${page} (${type})`
-      if (!endpoint || changeSubscriptions.has(key)) return
-
-      const changedPromise = endpoint[`${type}Changed`](includeIssues)
-      changeSubscriptions.set(key, changedPromise)
-      const changed = await changedPromise
-
-      for await (const change of changed) {
-        processIssues(page, change)
-        const payload = await makePayload(page, change)
-        if (payload) {
-          sendHmr('endpoint-change', key, payload)
-        }
-      }
-    }
-
-    async function clearChangeSubscription(
-      page: string,
-      type: 'server' | 'client'
-    ) {
-      const key = `${page} (${type})`
-      const subscription = await changeSubscriptions.get(key)
-      if (subscription) {
-        subscription.return?.()
-        changeSubscriptions.delete(key)
-      }
-      issues.delete(key)
-    }
-
-    async function writeBuildManifest(
-      rewrites: SetupOpts['fsChecker']['rewrites']
-    ): Promise<void> {
-      const buildManifest = mergeBuildManifests(buildManifests.values())
-      const buildManifestPath = path.join(distDir, BUILD_MANIFEST)
-      const middlewareBuildManifestPath = path.join(
-        distDir,
-        'server',
-        `${MIDDLEWARE_BUILD_MANIFEST}.js`
-      )
-      const interceptionRewriteManifestPath = path.join(
-        distDir,
-        'server',
-        `${INTERCEPTION_ROUTE_REWRITE_MANIFEST}.js`
-      )
-      deleteCache(buildManifestPath)
-      deleteCache(middlewareBuildManifestPath)
-      deleteCache(interceptionRewriteManifestPath)
-      await writeFileAtomic(
-        buildManifestPath,
-        JSON.stringify(buildManifest, null, 2)
-      )
-      await writeFileAtomic(
-        middlewareBuildManifestPath,
-        `self.__BUILD_MANIFEST=${JSON.stringify(buildManifest)};`
-      )
-
-      const interceptionRewrites = JSON.stringify(
-        rewrites.beforeFiles.filter(isInterceptionRouteRewrite)
-      )
-
-      await writeFileAtomic(
-        interceptionRewriteManifestPath,
-        `self.__INTERCEPTION_ROUTE_REWRITE_MANIFEST=${JSON.stringify(
-          interceptionRewrites
-        )};`
-      )
-
-      const content: ClientBuildManifest = {
-        __rewrites: rewrites
-          ? (normalizeRewritesForBuildManifest(rewrites) as any)
-          : { afterFiles: [], beforeFiles: [], fallback: [] },
-        ...Object.fromEntries(
-          [...curEntries.keys()].map((pathname) => [
-            pathname,
-            `static/chunks/pages${pathname === '/' ? '/index' : pathname}.js`,
-          ])
-        ),
-        sortedPages: [...curEntries.keys()],
-      }
-      const buildManifestJs = `self.__BUILD_MANIFEST = ${JSON.stringify(
-        content
-      )};self.__BUILD_MANIFEST_CB && self.__BUILD_MANIFEST_CB()`
-      await writeFileAtomic(
-        path.join(distDir, 'static', 'development', '_buildManifest.js'),
-        buildManifestJs
-      )
-      await writeFileAtomic(
-        path.join(distDir, 'static', 'development', '_ssgManifest.js'),
-        srcEmptySsgManifest
-      )
-    }
-
-    async function writeFallbackBuildManifest(): Promise<void> {
-      const fallbackBuildManifest = mergeBuildManifests(
-        [buildManifests.get('_app'), buildManifests.get('_error')].filter(
-          Boolean
-        ) as BuildManifest[]
-      )
-      const fallbackBuildManifestPath = path.join(
-        distDir,
-        `fallback-${BUILD_MANIFEST}`
-      )
-      deleteCache(fallbackBuildManifestPath)
-      await writeFileAtomic(
-        fallbackBuildManifestPath,
-        JSON.stringify(fallbackBuildManifest, null, 2)
-      )
-    }
-
-    async function writeAppBuildManifest(): Promise<void> {
-      const appBuildManifest = mergeAppBuildManifests(
-        appBuildManifests.values()
-      )
-      const appBuildManifestPath = path.join(distDir, APP_BUILD_MANIFEST)
-      deleteCache(appBuildManifestPath)
-      await writeFileAtomic(
-        appBuildManifestPath,
-        JSON.stringify(appBuildManifest, null, 2)
-      )
-    }
-
-    async function writePagesManifest(): Promise<void> {
-      const pagesManifest = mergePagesManifests(pagesManifests.values())
-      const pagesManifestPath = path.join(distDir, 'server', PAGES_MANIFEST)
-      deleteCache(pagesManifestPath)
-      await writeFileAtomic(
-        pagesManifestPath,
-        JSON.stringify(pagesManifest, null, 2)
-      )
-    }
-
-    async function writeAppPathsManifest(): Promise<void> {
-      const appPathsManifest = mergePagesManifests(appPathsManifests.values())
-      const appPathsManifestPath = path.join(
-        distDir,
-        'server',
-        APP_PATHS_MANIFEST
-      )
-      deleteCache(appPathsManifestPath)
-      await writeFileAtomic(
-        appPathsManifestPath,
-        JSON.stringify(appPathsManifest, null, 2)
-      )
-    }
-
-    async function writeMiddlewareManifest(): Promise<void> {
-      const middlewareManifest = mergeMiddlewareManifests(
-        middlewareManifests.values()
-      )
-      const middlewareManifestPath = path.join(
-        distDir,
-        'server',
-        MIDDLEWARE_MANIFEST
-      )
-      deleteCache(middlewareManifestPath)
-      await writeFileAtomic(
-        middlewareManifestPath,
-        JSON.stringify(middlewareManifest, null, 2)
-      )
-    }
-
-    async function writeActionManifest(): Promise<void> {
-      const actionManifest = await mergeActionManifests(
-        actionManifests.values()
-      )
-      const actionManifestJsonPath = path.join(
-        distDir,
-        'server',
-        `${SERVER_REFERENCE_MANIFEST}.json`
-      )
-      const actionManifestJsPath = path.join(
-        distDir,
-        'server',
-        `${SERVER_REFERENCE_MANIFEST}.js`
-      )
-      const json = JSON.stringify(actionManifest, null, 2)
-      deleteCache(actionManifestJsonPath)
-      deleteCache(actionManifestJsPath)
-      await writeFile(actionManifestJsonPath, json, 'utf-8')
-      await writeFile(
-        actionManifestJsPath,
-        `self.__RSC_SERVER_MANIFEST=${JSON.stringify(json)}`,
-        'utf-8'
-      )
-    }
-
-    async function writeFontManifest(): Promise<void> {
-      const fontManifest = mergeFontManifests(fontManifests.values())
-      const json = JSON.stringify(fontManifest, null, 2)
-
-      const fontManifestJsonPath = path.join(
-        distDir,
-        'server',
-        `${NEXT_FONT_MANIFEST}.json`
-      )
-      const fontManifestJsPath = path.join(
-        distDir,
-        'server',
-        `${NEXT_FONT_MANIFEST}.js`
-      )
-      deleteCache(fontManifestJsonPath)
-      deleteCache(fontManifestJsPath)
-      await writeFileAtomic(fontManifestJsonPath, json)
-      await writeFileAtomic(
-        fontManifestJsPath,
-        `self.__NEXT_FONT_MANIFEST=${JSON.stringify(json)}`
-      )
-    }
-
-    async function writeLoadableManifest(): Promise<void> {
-      const loadableManifest = mergeLoadableManifests(
-        loadableManifests.values()
-      )
-      const loadableManifestPath = path.join(distDir, REACT_LOADABLE_MANIFEST)
-      const middlewareloadableManifestPath = path.join(
-        distDir,
-        'server',
-        `${MIDDLEWARE_REACT_LOADABLE_MANIFEST}.js`
-      )
-
-      const json = JSON.stringify(loadableManifest, null, 2)
-
-      deleteCache(loadableManifestPath)
-      deleteCache(middlewareloadableManifestPath)
-      await writeFileAtomic(loadableManifestPath, json)
-      await writeFileAtomic(
-        middlewareloadableManifestPath,
-        `self.__REACT_LOADABLE_MANIFEST=${JSON.stringify(json)}`
-      )
-    }
-
-    async function writeManifests(): Promise<void> {
-      await writeBuildManifest(opts.fsChecker.rewrites)
-      await writeAppBuildManifest()
-      await writePagesManifest()
-      await writeAppPathsManifest()
-      await writeMiddlewareManifest()
-      await writeActionManifest()
-      await writeFontManifest()
-      await writeLoadableManifest()
-      await writeFallbackBuildManifest()
-    }
-
-    async function subscribeToHmrEvents(id: string, client: ws) {
-      let mapping = clientToHmrSubscription.get(client)
-      if (mapping === undefined) {
-        mapping = new Map()
-        clientToHmrSubscription.set(client, mapping)
-      }
-      if (mapping.has(id)) return
-
-      const subscription = project!.hmrEvents(id)
-      mapping.set(id, subscription)
-
-      // The subscription will always emit once, which is the initial
-      // computation. This is not a change, so swallow it.
-      try {
-        await subscription.next()
-
-        for await (const data of subscription) {
-          processIssues(id, data)
-          if (data.type !== 'issues') {
-            sendTurbopackMessage(data)
-          }
-        }
-      } catch (e) {
-        // The client might be using an HMR session from a previous server, tell them
-        // to fully reload the page to resolve the issue. We can't use
-        // `hotReloader.send` since that would force very connected client to
-        // reload, only this client is out of date.
-        const reloadAction: ReloadPageAction = {
-          action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE,
-        }
-        client.send(JSON.stringify(reloadAction))
-        client.close()
-        return
-      }
-    }
-
-    function unsubscribeToHmrEvents(id: string, client: ws) {
-      const mapping = clientToHmrSubscription.get(client)
-      const subscription = mapping?.get(id)
-      subscription?.return!()
-    }
-
-    try {
-      async function handleEntries() {
-        for await (const entrypoints of iter) {
-          if (!currentEntriesHandlingResolve) {
-            currentEntriesHandling = new Promise(
-              // eslint-disable-next-line no-loop-func
-              (resolve) => (currentEntriesHandlingResolve = resolve)
-            )
-          }
-          globalEntries.app = entrypoints.pagesAppEndpoint
-          globalEntries.document = entrypoints.pagesDocumentEndpoint
-          globalEntries.error = entrypoints.pagesErrorEndpoint
-
-          curEntries.clear()
-
-          for (const [pathname, route] of entrypoints.routes) {
-            switch (route.type) {
-              case 'page':
-              case 'page-api':
-              case 'app-page':
-              case 'app-route': {
-                curEntries.set(pathname, route)
-                break
-              }
-              default:
-                Log.info(`skipping ${pathname} (${route.type})`)
-                break
-            }
-          }
-
-          for (const [pathname, subscriptionPromise] of changeSubscriptions) {
-            if (pathname === '') {
-              // middleware is handled below
-              continue
-            }
-
-            if (!curEntries.has(pathname)) {
-              const subscription = await subscriptionPromise
-              subscription.return?.()
-              changeSubscriptions.delete(pathname)
-            }
-          }
-
-          const { middleware, instrumentation } = entrypoints
-          // We check for explicit true/false, since it's initialized to
-          // undefined during the first loop (middlewareChanges event is
-          // unnecessary during the first serve)
-          if (prevMiddleware === true && !middleware) {
-            // Went from middleware to no middleware
-            await clearChangeSubscription('middleware', 'server')
-            sendHmr('entrypoint-change', 'middleware', {
-              event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES,
-            })
-          } else if (prevMiddleware === false && middleware) {
-            // Went from no middleware to middleware
-            sendHmr('endpoint-change', 'middleware', {
-              event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES,
-            })
-          }
-          if (
-            opts.nextConfig.experimental.instrumentationHook &&
-            instrumentation
-          ) {
-            const processInstrumentation = async (
-              displayName: string,
-              name: string,
-              prop: 'nodeJs' | 'edge'
-            ) => {
-              const writtenEndpoint = await processResult(
-                displayName,
-                await instrumentation[prop].writeToDisk()
-              )
-              processIssues(name, writtenEndpoint)
-            }
-            await processInstrumentation(
-              'instrumentation (node.js)',
-              'instrumentation.nodeJs',
-              'nodeJs'
-            )
-            await processInstrumentation(
-              'instrumentation (edge)',
-              'instrumentation.edge',
-              'edge'
-            )
-            await loadMiddlewareManifest('instrumentation', 'instrumentation')
-            await writeManifests()
-
-            serverFields.actualInstrumentationHookFile = '/instrumentation'
-            await propagateServerField(
-              'actualInstrumentationHookFile',
-              serverFields.actualInstrumentationHookFile
-            )
-          } else {
-            serverFields.actualInstrumentationHookFile = undefined
-            await propagateServerField(
-              'actualInstrumentationHookFile',
-              serverFields.actualInstrumentationHookFile
-            )
-          }
-          if (middleware) {
-            const processMiddleware = async () => {
-              const writtenEndpoint = await processResult(
-                'middleware',
-                await middleware.endpoint.writeToDisk()
-              )
-              processIssues('middleware', writtenEndpoint)
-              await loadMiddlewareManifest('middleware', 'middleware')
-              serverFields.middleware = {
-                match: null as any,
-                page: '/',
-                matchers:
-                  middlewareManifests.get('middleware')?.middleware['/']
-                    .matchers,
-              }
-            }
-            await processMiddleware()
-
-            changeSubscription(
-              'middleware',
-              'server',
-              false,
-              middleware.endpoint,
-              async () => {
-                const finishBuilding = startBuilding(
-                  'middleware',
-                  undefined,
-                  true
-                )
-                await processMiddleware()
-                await propagateServerField(
-                  'actualMiddlewareFile',
-                  serverFields.actualMiddlewareFile
-                )
-                await propagateServerField(
-                  'middleware',
-                  serverFields.middleware
-                )
-                await writeManifests()
-
-                finishBuilding()
-                return { event: HMR_ACTIONS_SENT_TO_BROWSER.MIDDLEWARE_CHANGES }
-              }
-            )
-            prevMiddleware = true
-          } else {
-            middlewareManifests.delete('middleware')
-            serverFields.actualMiddlewareFile = undefined
-            serverFields.middleware = undefined
-            prevMiddleware = false
-          }
-          await propagateServerField(
-            'actualMiddlewareFile',
-            serverFields.actualMiddlewareFile
-          )
-          await propagateServerField('middleware', serverFields.middleware)
-
-          currentEntriesHandlingResolve!()
-          currentEntriesHandlingResolve = undefined
-        }
-      }
-
-      handleEntries().catch((err) => {
-        console.error(err)
-        process.exit(1)
-      })
-    } catch (e) {
-      console.error(e)
-    }
-
-    // Write empty manifests
-    await mkdir(path.join(distDir, 'server'), { recursive: true })
-    await mkdir(path.join(distDir, 'static/development'), { recursive: true })
-    await writeFile(
-      path.join(distDir, 'package.json'),
-      JSON.stringify(
-        {
-          type: 'commonjs',
-        },
-        null,
-        2
-      )
-    )
-    await currentEntriesHandling
-    await writeManifests()
-
-    const overlayMiddleware = getOverlayMiddleware(project)
-    let versionInfo: VersionInfo = await getVersionInfo(
-      true || isTestMode || opts.telemetry.isEnabled
-    )
-    const hotReloader: NextJsHotReloaderInterface = {
-      turbopackProject: project,
-      activeWebpackConfigs: undefined,
-      serverStats: null,
-      edgeServerStats: null,
-      async run(req, res, _parsedUrl) {
-        // intercept page chunks request and ensure them with turbopack
-        if (req.url?.startsWith('/_next/static/chunks/pages/')) {
-          const params = matchNextPageBundleRequest(req.url)
-
-          if (params) {
-            const decodedPagePath = `/${params.path
-              .map((param: string) => decodeURIComponent(param))
-              .join('/')}`
-
-            const denormalizedPagePath = denormalizePagePath(decodedPagePath)
-
-            await hotReloader
-              .ensurePage({
-                page: denormalizedPagePath,
-                clientOnly: false,
-                definition: undefined,
-                url: req.url,
-              })
-              .catch(console.error)
-          }
-        }
-
-        await overlayMiddleware(req, res)
-
-        // Request was not finished.
-        return { finished: undefined }
-      },
-
-      // TODO: Figure out if socket type can match the NextJsHotReloaderInterface
-      onHMR(req, socket: Socket, head) {
-        wsServer.handleUpgrade(req, socket, head, (client) => {
-          clients.add(client)
-          client.on('close', () => clients.delete(client))
-
-          client.addEventListener('message', ({ data }) => {
-            const parsedData = JSON.parse(
-              typeof data !== 'string' ? data.toString() : data
-            )
-
-            // Next.js messages
-            switch (parsedData.event) {
-              case 'ping':
-                // Ping doesn't need additional handling in Turbopack.
-                break
-              case 'span-end': {
-                hotReloaderSpan.manualTraceChild(
-                  parsedData.spanName,
-                  msToNs(parsedData.startTime),
-                  msToNs(parsedData.endTime),
-                  parsedData.attributes
-                )
-                break
-              }
-              case 'client-hmr-latency': // { id, startTime, endTime, page, updatedModules, isPageHidden }
-                hotReloaderSpan.manualTraceChild(
-                  parsedData.event,
-                  msToNs(parsedData.startTime),
-                  msToNs(parsedData.endTime),
-                  {
-                    updatedModules: parsedData.updatedModules,
-                    page: parsedData.page,
-                    isPageHidden: parsedData.isPageHidden,
-                  }
-                )
-                break
-              case 'client-error': // { errorCount, clientId }
-              case 'client-warning': // { warningCount, clientId }
-              case 'client-success': // { clientId }
-              case 'server-component-reload-page': // { clientId }
-              case 'client-reload-page': // { clientId }
-              case 'client-removed-page': // { page }
-              case 'client-full-reload': // { stackTrace, hadRuntimeError }
-              case 'client-added-page':
-                // TODO
-                break
-
-              default:
-                // Might be a Turbopack message...
-                if (!parsedData.type) {
-                  throw new Error(`unrecognized HMR message "${data}"`)
-                }
-            }
-
-            // Turbopack messages
-            switch (parsedData.type) {
-              case 'turbopack-subscribe':
-                subscribeToHmrEvents(parsedData.path, client)
-                break
-
-              case 'turbopack-unsubscribe':
-                unsubscribeToHmrEvents(parsedData.path, client)
-                break
-
-              default:
-                if (!parsedData.event) {
-                  throw new Error(
-                    `unrecognized Turbopack HMR message "${data}"`
-                  )
-                }
-            }
-          })
-
-          const turbopackConnected: TurbopackConnectedAction = {
-            action: HMR_ACTIONS_SENT_TO_BROWSER.TURBOPACK_CONNECTED,
-          }
-          client.send(JSON.stringify(turbopackConnected))
-
-          const errors = []
-          for (const pageIssues of issues.values()) {
-            for (const issue of pageIssues.values()) {
-              errors.push({
-                message: formatIssue(issue),
-              })
-            }
-          }
-
-          const sync: SyncAction = {
-            action: HMR_ACTIONS_SENT_TO_BROWSER.SYNC,
-            errors,
-            warnings: [],
-            hash: '',
-            versionInfo,
-          }
-
-          this.send(sync)
-        })
-      },
-
-      send(action) {
-        const payload = JSON.stringify(action)
-        for (const client of clients) {
-          client.send(payload)
-        }
-      },
-
-      setHmrServerError(_error) {
-        // Not implemented yet.
-      },
-      clearHmrServerError() {
-        // Not implemented yet.
-      },
-      async start() {
-        const enabled = isTestMode || opts.telemetry.isEnabled
-        const nextVersionInfo = await getVersionInfo(enabled)
-        if (nextVersionInfo) {
-          versionInfo = nextVersionInfo
-        }
-      },
-      async stop() {
-        // Not implemented yet.
-      },
-      async getCompilationErrors(page) {
-        const thisPageIssues = issues.get(page)
-        if (thisPageIssues !== undefined && thisPageIssues.size > 0) {
-          // If there is an error related to the requesting page we display it instead of the first error
-          return [...thisPageIssues.values()].map(
-            (issue) => new Error(formatIssue(issue))
-          )
-        }
-
-        // Otherwise, return all errors across pages
-        const errors = []
-        for (const pageIssues of issues.values()) {
-          for (const issue of pageIssues.values()) {
-            errors.push(new Error(formatIssue(issue)))
-          }
-        }
-        return errors
-      },
-      async invalidate({
-        // .env files or tsconfig/jsconfig change
-        reloadAfterInvalidation,
-      }) {
-        if (reloadAfterInvalidation) {
-          await clearAllModuleContexts()
-          this.send({
-            action: HMR_ACTIONS_SENT_TO_BROWSER.SERVER_COMPONENT_CHANGES,
-          })
-        }
-      },
-      async buildFallbackError() {
-        // Not implemented yet.
-      },
-      async ensurePage({
-        page: inputPage,
-        // Unused parameters
-        // clientOnly,
-        // appPaths,
-        definition,
-        isApp,
-        url: requestUrl,
-      }) {
-        let page = definition?.pathname ?? inputPage
-
-        if (page === '/_error') {
-          let finishBuilding = startBuilding(page, requestUrl)
-          try {
-            if (globalEntries.app) {
-              const writtenEndpoint = await processResult(
-                '_app',
-                await globalEntries.app.writeToDisk()
-              )
-              processIssues('_app', writtenEndpoint)
-            }
-            await loadBuildManifest('_app')
-            await loadPagesManifest('_app')
-            await loadFontManifest('_app')
-
-            if (globalEntries.document) {
-              const writtenEndpoint = await processResult(
-                '_document',
-                await globalEntries.document.writeToDisk()
-              )
-              changeSubscription(
-                '_document',
-                'server',
-                false,
-                globalEntries.document,
-                () => {
-                  return { action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE }
-                }
-              )
-              processIssues('_document', writtenEndpoint)
-            }
-            await loadPagesManifest('_document')
-
-            if (globalEntries.error) {
-              const writtenEndpoint = await processResult(
-                '_error',
-                await globalEntries.error.writeToDisk()
-              )
-              processIssues(page, writtenEndpoint)
-            }
-            await loadBuildManifest('_error')
-            await loadPagesManifest('_error')
-            await loadFontManifest('_error')
-
-            await writeManifests()
-          } finally {
-            finishBuilding()
-          }
-          return
-        }
-        await currentEntriesHandling
-        const route =
-          curEntries.get(page) ??
-          curEntries.get(
-            normalizeAppPath(
-              normalizeMetadataRoute(definition?.page ?? inputPage)
-            )
-          )
-
-        if (!route) {
-          // TODO: why is this entry missing in turbopack?
-          if (page === '/_app') return
-          if (page === '/_document') return
-          if (page === '/middleware') return
-          if (page === '/src/middleware') return
-          if (page === '/instrumentation') return
-          if (page === '/src/instrumentation') return
-
-          throw new PageNotFoundError(`route not found ${page}`)
-        }
-
-        let finishBuilding: (() => void) | undefined = undefined
-
-        try {
-          switch (route.type) {
-            case 'page': {
-              if (isApp) {
-                throw new Error(
-                  `mis-matched route type: isApp && page for ${page}`
-                )
-              }
-
-              finishBuilding = startBuilding(page, requestUrl)
-              try {
-                if (globalEntries.app) {
-                  const writtenEndpoint = await processResult(
-                    '_app',
-                    await globalEntries.app.writeToDisk()
-                  )
-                  processIssues('_app', writtenEndpoint)
-                }
-                await loadBuildManifest('_app')
-                await loadPagesManifest('_app')
-
-                if (globalEntries.document) {
-                  const writtenEndpoint = await processResult(
-                    '_document',
-                    await globalEntries.document.writeToDisk()
-                  )
-                  processIssues('_document', writtenEndpoint)
-                }
-                await loadPagesManifest('_document')
-
-                const writtenEndpoint = await processResult(
-                  page,
-                  await route.htmlEndpoint.writeToDisk()
-                )
-
-                const type = writtenEndpoint?.type
-
-                await loadBuildManifest(page)
-                await loadPagesManifest(page)
-                if (type === 'edge') {
-                  await loadMiddlewareManifest(page, 'pages')
-                } else {
-                  middlewareManifests.delete(page)
-                }
-                await loadFontManifest(page, 'pages')
-                await loadLoadableManifest(page, 'pages')
-
-                await writeManifests()
-
-                processIssues(page, writtenEndpoint)
-              } finally {
-                changeSubscription(
-                  page,
-                  'server',
-                  false,
-                  route.dataEndpoint,
-                  (pageName) => {
-                    // Report the next compilation again
-                    readyIds.delete(page)
-                    return {
-                      event: HMR_ACTIONS_SENT_TO_BROWSER.SERVER_ONLY_CHANGES,
-                      pages: [pageName],
-                    }
-                  }
-                )
-                changeSubscription(
-                  page,
-                  'client',
-                  false,
-                  route.htmlEndpoint,
-                  () => {
-                    return {
-                      event: HMR_ACTIONS_SENT_TO_BROWSER.CLIENT_CHANGES,
-                    }
-                  }
-                )
-                if (globalEntries.document) {
-                  changeSubscription(
-                    '_document',
-                    'server',
-                    false,
-                    globalEntries.document,
-                    () => {
-                      return { action: HMR_ACTIONS_SENT_TO_BROWSER.RELOAD_PAGE }
-                    }
-                  )
-                }
-              }
-
-              break
-            }
-            case 'page-api': {
-              // We don't throw on ensureOpts.isApp === true here
-              // since this can happen when app pages make
-              // api requests to page API routes.
-
-              finishBuilding = startBuilding(page, requestUrl)
-              const writtenEndpoint = await processResult(
-                page,
-                await route.endpoint.writeToDisk()
-              )
-
-              const type = writtenEndpoint?.type
-
-              await loadPagesManifest(page)
-              if (type === 'edge') {
-                await loadMiddlewareManifest(page, 'pages')
-              } else {
-                middlewareManifests.delete(page)
-              }
-              await loadLoadableManifest(page, 'pages')
-
-              await writeManifests()
-
-              processIssues(page, writtenEndpoint)
-
-              break
-            }
-            case 'app-page': {
-              finishBuilding = startBuilding(page, requestUrl)
-              const writtenEndpoint = await processResult(
-                page,
-                await route.htmlEndpoint.writeToDisk()
-              )
-
-              changeSubscription(
-                page,
-                'server',
-                true,
-                route.rscEndpoint,
-                (_page, change) => {
-                  if (
-                    change.issues.some((issue) => issue.severity === 'error')
-                  ) {
-                    // Ignore any updates that has errors
-                    // There will be another update without errors eventually
-                    return
-                  }
-                  // Report the next compilation again
-                  readyIds.delete(page)
-                  return {
-                    action:
-                      HMR_ACTIONS_SENT_TO_BROWSER.SERVER_COMPONENT_CHANGES,
-                  }
-                }
-              )
-
-              const type = writtenEndpoint?.type
-
-              if (type === 'edge') {
-                await loadMiddlewareManifest(page, 'app')
-              } else {
-                middlewareManifests.delete(page)
-              }
-
-              await loadAppBuildManifest(page)
-              await loadBuildManifest(page, 'app')
-              await loadAppPathManifest(page, 'app')
-              await loadActionManifest(page)
-              await loadFontManifest(page, 'app')
-              await writeManifests()
-
-              processIssues(page, writtenEndpoint, true)
-
-              break
-            }
-            case 'app-route': {
-              finishBuilding = startBuilding(page, requestUrl)
-              const writtenEndpoint = await processResult(
-                page,
-                await route.endpoint.writeToDisk()
-              )
-
-              const type = writtenEndpoint?.type
-
-              await loadAppPathManifest(page, 'app-route')
-              if (type === 'edge') {
-                await loadMiddlewareManifest(page, 'app-route')
-              } else {
-                middlewareManifests.delete(page)
-              }
-
-              await writeManifests()
-
-              processIssues(page, writtenEndpoint, true)
-
-              break
-            }
-            default: {
-              throw new Error(
-                `unknown route type ${(route as any).type} for ${page}`
-              )
-            }
-          }
-        } finally {
-          if (finishBuilding) finishBuilding()
-        }
-      },
-    }
-
-    ;(async function () {
-      for await (const updateMessage of project.updateInfoSubscribe(30)) {
-        switch (updateMessage.updateType) {
-          case 'start': {
-            hotReloader.send({ action: HMR_ACTIONS_SENT_TO_BROWSER.BUILDING })
-            break
-          }
-          case 'end': {
-            sendEnqueuedMessages()
-
-            const errors = new Map<string, CompilationError>()
-            for (const [, issueMap] of issues) {
-              for (const [key, issue] of issueMap) {
-                if (errors.has(key)) continue
-
-                const message = formatIssue(issue)
-
-                errors.set(key, {
-                  message,
-                  details: issue.detail
-                    ? renderStyledStringToErrorAnsi(issue.detail)
-                    : undefined,
-                })
-              }
-            }
-
-            hotReloader.send({
-              action: HMR_ACTIONS_SENT_TO_BROWSER.BUILT,
-              hash: String(++hmrHash),
-              errors: [...errors.values()],
-              warnings: [],
-            })
-
-            if (hmrEventHappened) {
-              const time = updateMessage.value.duration
-              const timeMessage =
-                time > 2000 ? `${Math.round(time / 100) / 10}s` : `${time}ms`
-              Log.event(`Compiled in ${timeMessage}`)
-              hmrEventHappened = false
-            }
-            break
-          }
-          default:
-        }
-      }
-    })()
-
-    return hotReloader
-  }
+  const serverFields: ServerFields = {}
 
   const hotReloader: NextJsHotReloaderInterface = opts.turbo
-    ? await createHotReloaderTurbopack()
+    ? await createHotReloaderTurbopack(opts, serverFields, distDir)
     : new HotReloaderWebpack(opts.dir, {
         appDir,
         pagesDir,
@@ -1884,6 +352,7 @@ async function startWatcher(opts: SetupOpts) {
           }
           serverFields.actualMiddlewareFile = rootFile
           await propagateServerField(
+            opts,
             'actualMiddlewareFile',
             serverFields.actualMiddlewareFile
           )
@@ -1898,6 +367,7 @@ async function startWatcher(opts: SetupOpts) {
         ) {
           serverFields.actualInstrumentationHookFile = rootFile
           await propagateServerField(
+            opts,
             'actualInstrumentationHookFile',
             serverFields.actualInstrumentationHookFile
           )
@@ -2010,7 +480,7 @@ async function startWatcher(opts: SetupOpts) {
           hotReloader.setHmrServerError(new Error(errorMessage))
         } else if (numConflicting === 0) {
           hotReloader.clearHmrServerError()
-          await propagateServerField('reloadMatchers', undefined)
+          await propagateServerField(opts, 'reloadMatchers', undefined)
         }
       }
 
@@ -2054,7 +524,7 @@ async function startWatcher(opts: SetupOpts) {
           loadEnvConfig(dir, true, Log, true, (envFilePath) => {
             Log.info(`Reload env: ${envFilePath}`)
           })
-          await propagateServerField('loadEnvConfig', [
+          await propagateServerField(opts, 'loadEnvConfig', [
             { dev: true, forceReload: true, silent: true },
           ])
         }
@@ -2196,7 +666,11 @@ async function startWatcher(opts: SetupOpts) {
       serverFields.appPathRoutes = Object.fromEntries(
         Object.entries(appPaths).map(([k, v]) => [k, v.sort()])
       )
-      await propagateServerField('appPathRoutes', serverFields.appPathRoutes)
+      await propagateServerField(
+        opts,
+        'appPathRoutes',
+        serverFields.appPathRoutes
+      )
 
       // TODO: pass this to fsChecker/next-dev-server?
       serverFields.middleware = middlewareMatchers
@@ -2207,7 +681,7 @@ async function startWatcher(opts: SetupOpts) {
           }
         : undefined
 
-      await propagateServerField('middleware', serverFields.middleware)
+      await propagateServerField(opts, 'middleware', serverFields.middleware)
       serverFields.hasAppNotFound = hasRootAppNotFound
 
       opts.fsChecker.middlewareMatcher = serverFields.middleware?.matchers
@@ -2351,7 +825,7 @@ async function startWatcher(opts: SetupOpts) {
       } finally {
         // Reload the matchers. The filesystem would have been written to,
         // and the matchers need to re-scan it to update the router.
-        await propagateServerField('reloadMatchers', undefined)
+        await propagateServerField(opts, 'reloadMatchers', undefined)
       }
     })
 
@@ -2559,34 +1033,3 @@ export async function setupDevBundler(opts: SetupOpts) {
 }
 
 export type DevBundler = Awaited<ReturnType<typeof setupDevBundler>>
-
-function renderStyledStringToErrorAnsi(string: StyledString): string {
-  function decodeMagicIdentifiers(str: string): string {
-    return str.replaceAll(MAGIC_IDENTIFIER_REGEX, (ident) => {
-      try {
-        return magenta(`{${decodeMagicIdentifier(ident)}}`)
-      } catch (e) {
-        return magenta(`{${ident} (decoding failed: ${e})}`)
-      }
-    })
-  }
-
-  switch (string.type) {
-    case 'text':
-      return decodeMagicIdentifiers(string.value)
-    case 'strong':
-      return bold(red(decodeMagicIdentifiers(string.value)))
-    case 'code':
-      return green(decodeMagicIdentifiers(string.value))
-    case 'line':
-      return string.value.map(renderStyledStringToErrorAnsi).join('')
-    case 'stack':
-      return string.value.map(renderStyledStringToErrorAnsi).join('\n')
-    default:
-      throw new Error('Unknown StyledString type', string)
-  }
-}
-
-function msToNs(ms: number): bigint {
-  return BigInt(Math.floor(ms)) * BigInt(MILLISECONDS_IN_NANOSECOND)
-}
